Add ModPowerBonus to all weapons
Based on in-game stats, seems like the mod bonuses are not being applied to weapons
that are extended from another weapon, even though the base weapon
has those properties. It seems to be unintended, since the base weapons
have those properties. Therefore I added those properties to all
extended weapons. This causes the weapon's entity damage display in the
inventory to be correctly increased as mods are added.